+++ Tsk.HttpApi/Products/ForCustomers/DataTransferObjects.cs
@@ -7,5 +7,5 @@ public class ProductDto
 {
     public required Guid Id { get; init; }
     public required string Title { get; init; }
-    public required double Price { get; init; }
+    public required decimal Price { get; init; }
 }
